Moved uninit to own test file
