+++ test/init_test.py
@@ -1,5 +1,5 @@
 import os
-from lamby.lamby import init, uninit
+from lamby.lamby import init
 
 
 def test_init(runner):
@@ -19,12 +19,3 @@ def test_init(runner):
         with open('./.lamby/log', 'r') as log_file:
             data = log_file.read()
             assert data == '{}'
-
-
-def test_uninit(runner):
-    with runner.isolated_filesystem():
-        result = runner.invoke(init, [])
-        result = runner.invoke(uninit, [])
-
-        assert result.exit_code == 0
-        assert not os.path.isdir('./.lamby')
